refactor: editor components composition
