+++ src/modules/domain/rich-text/prosemirror/notes/plugin.ts
@@ -5,6 +5,7 @@ import { noteContentNumbering } from '../../../../../renderer/src/components/edi
 import {
   createNoteNumberingTransaction,
   ensureTrailingSpaceAfterContentBlockNumbers,
+  numberNotes,
 } from './commands';
 import { handleBackspaceOrDelete } from './events';
 import { getNotes } from './state';
@@ -14,6 +15,22 @@ const pluginKey = new PluginKey('note-numbering');
 export const notesPlugin = () =>
   new Plugin({
     key: pluginKey,
+    // Initial numbering pass: stored numbering may be stale, so dispatch one renumbering
+    // transaction up-front. Deferred with queueMicrotask: plugin views are constructed
+    // inside EditorView creation/updateState, and a synchronous dispatch would start
+    // a second update while the first is still on the stack.
+    view(editorView) {
+      let cancelled = false;
+      queueMicrotask(() => {
+        if (cancelled || editorView.isDestroyed) return;
+        numberNotes(editorView.state, editorView.dispatch);
+      });
+      return {
+        destroy() {
+          cancelled = true;
+        },
+      };
+    },
     state: {
       init: () => ({}),
       apply(_, prev) {
