--- src/renderer/src/components/editing/RichTextEditor.tsx
@@ -1,733 +1,0 @@
-import { clsx } from 'clsx';
-import debounce from 'debounce';
-import * as Effect from 'effect/Effect';
-import * as SubscriptionRef from 'effect/SubscriptionRef';
-import {
-  baseKeymap,
-  chainCommands,
-  setBlockType as setProsemirrorBlockType,
-} from 'prosemirror-commands';
-import { history, redo, undo } from 'prosemirror-history';
-import { keymap } from 'prosemirror-keymap';
-import { type Node, type Schema } from 'prosemirror-model';
-import { EditorState, Selection, Transaction } from 'prosemirror-state';
-import { EditorView } from 'prosemirror-view';
-import { useCallback, useContext, useEffect, useRef, useState } from 'react';
-
-import {
-  type BlockType,
-  blockTypes,
-  type ContainerBlockType,
-  CURRENT_SCHEMA_VERSION,
-  getDocumentRichTextContent,
-  getHeadingLevel,
-  type LeafBlockType,
-  LinkAttrs,
-  type LiveDocument,
-  prosemirror,
-  type RichTextDocument,
-  richTextRepresentations,
-} from '../../../../modules/domain/rich-text';
-import { ProseMirrorContext } from '../../../../modules/domain/rich-text/react/prosemirror-context';
-import {
-  createErrorNotification,
-  NotificationsContext,
-} from '../../../../modules/infrastructure/notifications/browser';
-import { useKeyBindings } from '../../keyboard';
-import { keyBindings } from '../../pages/project/shared/command-palette/key-bindings';
-import { EditorToolbar } from './editor-toolbar';
-import { LinkDialog } from './LinkDialog';
-import { LinkPopover } from './LinkPopover';
-
-const {
-  schema,
-  buildInputRules,
-  getCurrentLeafBlockType,
-  getCurrentContainerBlockType,
-  isMarkActive,
-  toggleEm,
-  toggleStrong,
-  toggleCode,
-  transactionUpdatesMarks,
-  addLink,
-  removeLink,
-  updateLink,
-  linkSelectionPlugin,
-  selectionChangePlugin,
-  getSelectedText,
-  findLinkAtSelection,
-  ensureTrailingParagraphInDoc,
-  ensureTrailingParagraphPlugin,
-  ensureTrailingSpaceAfterAtomPlugin,
-  moveCursorToNextBlockOnInsertionPlugin,
-  removeEmptyFiguresPlugin,
-  wrapInList,
-  wrapIn,
-  splitListItem,
-  liftListItem,
-  sinkListItem,
-  pasteMarkdownPlugin,
-  markdownMarkPlugins,
-  insertNote,
-  notesPlugin,
-  insertHorizontalRule,
-  canInsertHorizontalRule,
-  canInsertFigure,
-  deleteFigureBeforeCursor,
-  deleteFigureAfterCursor,
-  moveToParagraphAfterSelectedFigure,
-  pickAndInsertFigure,
-  numberNotes,
-  placeholderPlugin,
-  oneWaySyncPlugin,
-  liveSyncPlugin,
-  pmDocFromJSONString,
-  pmDocToJSONString,
-  diffPlugin,
-  registerNodeViews,
-  codeBlockHighlightPlugin,
-} = prosemirror;
-
-type SharedRichTextEditorProps = {
-  isToolbarOpen?: boolean;
-  pickAsset: prosemirror.FigureAssetPicker;
-  resolveAssetSrc: prosemirror.ResolveAssetSrc;
-};
-
-// Backed by a live document: always editable, content owned outside the editor.
-type LiveBackedProps = SharedRichTextEditorProps & {
-  live: LiveDocument;
-};
-
-// Backed by a plain value: the editor renders what it is given and reports
-// changes back through `onDocChange`.
-type ValueBackedProps = SharedRichTextEditorProps & {
-  doc: RichTextDocument;
-  onDocChange?: (doc: RichTextDocument) => Promise<void>;
-  isEditable?: boolean;
-  showDiffWith?: RichTextDocument;
-};
-
-type RichTextEditorProps = LiveBackedProps | ValueBackedProps;
-
-export const RichTextEditor = (props: RichTextEditorProps) => {
-  const { isToolbarOpen = false, pickAsset, resolveAssetSrc } = props;
-  const live = 'live' in props ? props.live : null;
-  const valueSource: ValueBackedProps | null = 'live' in props ? null : props;
-  const isEditable = valueSource ? (valueSource.isEditable ?? true) : true;
-  const showDiffWith = valueSource?.showDiffWith;
-
-  const editorRoot = useRef<HTMLDivElement>(null);
-  const editorViewRef = useRef<EditorView | null>(null);
-  const { dispatchNotification } = useContext(NotificationsContext);
-  const {
-    view,
-    setView,
-    parseMarkdown,
-    convertToProseMirror,
-    convertFromProseMirror,
-    proseMirrorDiff,
-  } = useContext(ProseMirrorContext);
-  const [leafBlockType, setLeafBlockType] = useState<LeafBlockType | null>(
-    null
-  );
-  const [containerBlockType, setContainerBlockType] =
-    useState<ContainerBlockType | null>(null);
-  const [strongSelected, setStrongSelected] = useState<boolean>(false);
-  const [codeSelected, setCodeSelected] = useState<boolean>(false);
-  const [emSelected, setEmSelected] = useState<boolean>(false);
-  const [selectionIsLink, setSelectionIsLink] = useState<boolean>(false);
-  const [horizontalRuleEnabled, setHorizontalRuleEnabled] =
-    useState<boolean>(false);
-  const [imageEnabled, setImageEnabled] = useState<boolean>(false);
-  const [isLinkDialogOpen, setIsLinkDialogOpen] = useState<boolean>(false);
-  const [linkDialogInitialAttrs, setLinkDialogInitialAttrs] =
-    useState<LinkAttrs>({ title: '', href: '' });
-  const [isLinkPopoverOpen, setIsLinkPopoverOpen] = useState<boolean>(false);
-  // Using state instead of useRef to trigger a popover re-render when the link ref changes
-  const [selectedLinkData, setSelectedLinkData] = useState<{
-    ref: Element;
-    linkAttrs: LinkAttrs;
-  } | null>(null);
-
-  const onSelectionChange: (
-    schema: Schema
-  ) => (selection: Selection, view: EditorView) => void =
-    (schema) => (selection, view) => {
-      const hideLinkPopover = () => {
-        setSelectedLinkData(null);
-        setIsLinkPopoverOpen(false);
-      };
-
-      if (isMarkActive(schema.marks.link)(view.state)) {
-        const link = findLinkAtSelection({ view, selection });
-        if (link) {
-          setSelectedLinkData({ ref: link.element, linkAttrs: link.linkAttrs });
-          setIsLinkPopoverOpen(true);
-        } else {
-          hideLinkPopover();
-        }
-      } else {
-        hideLinkPopover();
-      }
-    };
-
-  const getBasePlugins = (schema: Schema) => [
-    buildInputRules(schema),
-    placeholderPlugin('Start writing...'),
-    ...markdownMarkPlugins(schema),
-    pasteMarkdownPlugin(parseMarkdown(schema)),
-    // Must come before `notesPlugin`, which installs a `handleKeyDown`
-    // for Backspace/Delete that falls back to PM's default chain.
-    keymap({
-      Backspace: deleteFigureBeforeCursor,
-      Delete: deleteFigureAfterCursor,
-      Enter: chainCommands(
-        moveToParagraphAfterSelectedFigure,
-        splitListItem(schema.nodes.list_item)
-      ),
-      'Mod-b': toggleStrong(schema),
-      'Mod-i': toggleEm(schema),
-      'Mod-z': undo,
-      'Mod-y': redo,
-      'Shift-Mod-z': redo,
-      'Mod-[': liftListItem(schema.nodes.list_item),
-      'Mod-]': sinkListItem(schema.nodes.list_item),
-      'Mod-Alt-f': insertNote,
-      // Disable tab keystrokes in the editor to prevent tabbing
-      // to the next focusable element
-      Tab: () => true,
-    }),
-    notesPlugin(),
-    codeBlockHighlightPlugin,
-    history(),
-    keymap(baseKeymap),
-    linkSelectionPlugin,
-    selectionChangePlugin(onSelectionChange(schema)),
-    ensureTrailingParagraphPlugin(schema),
-    moveCursorToNextBlockOnInsertionPlugin(schema),
-    ensureTrailingSpaceAfterAtomPlugin(),
-    removeEmptyFiguresPlugin(schema),
-  ];
-
-  useKeyBindings({
-    [keyBindings.ctrlShiftL.keyBinding]: () => {
-      handleLinkToggle();
-    },
-  });
-
-  const setupDiffPlugin = async ({
-    currentDoc,
-    diffWith,
-  }: {
-    currentDoc: RichTextDocument;
-    diffWith: RichTextDocument;
-  }) => {
-    const contentBefore = getDocumentRichTextContent(diffWith);
-    const contentAfter = getDocumentRichTextContent(currentDoc);
-
-    const { decorations } = await proseMirrorDiff({
-      representation:
-        // There are some old document versions without the representataion set. The representation is Automerge in that case.
-        // TODO: Remove this fallback when we no longer expect documents without representation set.
-        valueSource?.doc.representation ?? richTextRepresentations.AUTOMERGE,
-      proseMirrorSchema: schema,
-      docBefore: contentBefore,
-      docAfter: contentAfter,
-      transformImageSrc: resolveAssetSrc,
-    });
-
-    return diffPlugin({
-      decorations,
-      proseMirrorDiff,
-      convertFromProseMirror,
-      transformImageSrc: resolveAssetSrc,
-      diffWith: showDiffWith,
-    });
-  };
-
-  const setupSyncPlugin = ({
-    doc,
-    onDocChange,
-  }: {
-    doc: RichTextDocument;
-    onDocChange: (doc: RichTextDocument) => Promise<void>;
-  }) => {
-    const handlePMDocChange = debounce(async (pmDoc: Node) => {
-      const pmJSONStr = pmDocToJSONString(pmDoc);
-
-      onDocChange({
-        schemaVersion: doc.schemaVersion,
-        representation: richTextRepresentations.PROSEMIRROR,
-        content: pmJSONStr,
-      });
-    }, 300);
-
-    return oneWaySyncPlugin({
-      onPMDocChange: handlePMDocChange,
-    });
-  };
-
-  const handleLiveSyncError = useCallback(
-    (error: unknown) => {
-      console.error(error);
-      dispatchNotification(
-        createErrorNotification({
-          title: 'Editor Sync Error',
-          message:
-            'An error happened while syncing this document. Please reach out to us for support.',
-        })
-      );
-    },
-    [dispatchNotification]
-  );
-
-  const setupLiveSyncPlugin = ({
-    live,
-    schema,
-    initial,
-  }: {
-    live: LiveDocument;
-    schema: Schema;
-    initial: { doc: RichTextDocument; version: string };
-  }) =>
-    liveSyncPlugin({
-      live,
-      initialVersion: initial.version,
-      schemaVersion: initial.doc.schemaVersion,
-      schema,
-      convertToProseMirror: (document) =>
-        convertToProseMirror({
-          schema,
-          document: {
-            ...document,
-            content: getDocumentRichTextContent(document),
-          },
-        }),
-      onError: handleLiveSyncError,
-    });
-
-  useEffect(() => {
-    let cancelled = false;
-
-    const buildLiveBackedEditor = async (
-      schema: Schema,
-      live: LiveDocument
-    ) => {
-      const plugins = getBasePlugins(schema);
-      const initial = Effect.runSync(SubscriptionRef.get(live.content));
-
-      plugins.push(setupLiveSyncPlugin({ live, schema, initial }));
-
-      const pmDoc =
-        initial.doc.representation === richTextRepresentations.PROSEMIRROR
-          ? pmDocFromJSONString(JSON.parse(initial.doc.content), schema)
-          : await convertToProseMirror({
-              schema: schema,
-              document: {
-                ...initial.doc,
-                content: getDocumentRichTextContent(initial.doc),
-              },
-            });
-
-      return { plugins, pmDoc };
-    };
-
-    const buildValueBackedEditor = async (
-      schema: Schema,
-      source: ValueBackedProps
-    ) => {
-      const { doc, onDocChange, showDiffWith } = source;
-      const plugins = getBasePlugins(schema);
-
-      if (isEditable && onDocChange) {
-        const sync = setupSyncPlugin({ doc, onDocChange });
-        plugins.push(sync);
-      }
-
-      if (showDiffWith) {
-        const diff = await setupDiffPlugin({
-          currentDoc: doc,
-          diffWith: showDiffWith,
-        });
-        plugins.push(diff);
-      }
-
-      const richTextContent = getDocumentRichTextContent(doc);
-
-      const pmDoc =
-        doc.representation !== richTextRepresentations.PROSEMIRROR
-          ? await convertToProseMirror({
-              schema: schema,
-              document: {
-                ...doc,
-                content: richTextContent,
-              },
-            })
-          : pmDocFromJSONString(doc.content, schema);
-
-      return { plugins, pmDoc };
-    };
-
-    const setupEditorAndView = async (schema: Schema) => {
-      const { plugins, pmDoc } =
-        'live' in props
-          ? await buildLiveBackedEditor(schema, props.live)
-          : await buildValueBackedEditor(schema, props);
-
-      // After awaiting async conversion, ensure this run still represents the
-      // current document and that another effect didn't create the view in the
-      // meantime (avoid duplicate EditorView creation).
-      if (cancelled || editorViewRef.current) return;
-
-      const editorConfig = {
-        schema,
-        plugins,
-        // Apply the trailing-paragraph invariant up-front so docs that end in a figure
-        // (or other block that needs one) get a trailing paragraph on first load.
-        // The relevant plugin fires after the first transaction, so it's not enough.
-        doc: ensureTrailingParagraphInDoc(pmDoc, schema),
-      };
-
-      const state = EditorState.create(editorConfig);
-      const editorView = new EditorView(editorRoot.current, {
-        state,
-        nodeViews: registerNodeViews({ resolveAssetSrc }),
-        dispatchTransaction: (tx: Transaction) => {
-          const newState = editorView.state.apply(tx);
-          editorView.updateState(newState);
-
-          // React state updates
-          setLeafBlockType(getCurrentLeafBlockType(newState));
-          setContainerBlockType(getCurrentContainerBlockType(newState));
-          setHorizontalRuleEnabled(canInsertHorizontalRule(newState));
-          setImageEnabled(canInsertFigure(newState));
-
-          if (tx.selectionSet || transactionUpdatesMarks(tx)) {
-            setStrongSelected(isMarkActive(schema.marks.strong)(newState));
-            setEmSelected(isMarkActive(schema.marks.em)(newState));
-            setSelectionIsLink(isMarkActive(schema.marks.link)(newState));
-            setCodeSelected(isMarkActive(schema.marks.code)(newState));
-          }
-        },
-        editable: () => isEditable,
-      });
-
-      editorViewRef.current = editorView;
-
-      numberNotes(state, editorView.dispatch, editorView);
-
-      // Announce the view to the shared context only after creation.
-      setView(editorView);
-
-      if (isEditable) {
-        editorViewRef.current?.focus();
-        setLeafBlockType(getCurrentLeafBlockType(state));
-        setContainerBlockType(getCurrentContainerBlockType(state));
-        setHorizontalRuleEnabled(canInsertHorizontalRule(state));
-        setImageEnabled(canInsertFigure(state));
-      }
-    };
-
-    if (schema && !editorViewRef.current) {
-      setupEditorAndView(schema);
-    }
-
-    return () => {
-      cancelled = true;
-
-      // If this component created the view (editorViewRef), destroy it and
-      // clear the context view so other components won't reuse the destroyed
-      // instance.
-      if (editorViewRef.current) {
-        editorViewRef.current.destroy();
-        // If the context still holds the same view, clear it.
-        if (view === editorViewRef.current) {
-          setView(null);
-        }
-        editorViewRef.current = null;
-      }
-    };
-  }, [valueSource?.doc, live, isEditable, schema, setView]);
-
-  // Only value-backed mode reconfigures plugins; the live-backed editor keeps
-  // the plugin set it was created with.
-  useEffect(() => {
-    const reconfigurePlugins = async ({
-      pmDoc,
-      diffWith,
-      editorView,
-      source,
-    }: {
-      pmDoc: Node;
-      diffWith: RichTextDocument | undefined;
-      editorView: EditorView;
-      source: ValueBackedProps;
-    }) => {
-      const plugins = getBasePlugins(schema);
-
-      if (isEditable && source.onDocChange) {
-        const sync = setupSyncPlugin({
-          doc: source.doc,
-          onDocChange: source.onDocChange,
-        });
-        plugins.push(sync);
-      }
-
-      if (diffWith) {
-        const currentDocContent = await convertFromProseMirror({
-          pmDoc,
-          to: diffWith.representation,
-        });
-
-        const currentDoc: RichTextDocument = {
-          representation: diffWith.representation,
-          schemaVersion: CURRENT_SCHEMA_VERSION,
-          content: currentDocContent,
-        };
-
-        const diffPlugin = await setupDiffPlugin({
-          currentDoc,
-          diffWith,
-        });
-
-        plugins.push(diffPlugin);
-      }
-
-      const newState = editorView.state.reconfigure({ plugins });
-      editorView.updateState(newState);
-    };
-
-    if (valueSource && editorViewRef.current) {
-      reconfigurePlugins({
-        pmDoc: editorViewRef.current.state.doc,
-        diffWith: showDiffWith,
-        editorView: editorViewRef.current,
-        source: valueSource,
-      });
-    }
-  }, [showDiffWith, convertFromProseMirror]);
-
-  const handleBlockSelect = (type: BlockType) => {
-    if (view) {
-      const { $from } = view.state.selection;
-
-      switch (type) {
-        case blockTypes.HEADING_1:
-        case blockTypes.HEADING_2:
-        case blockTypes.HEADING_3:
-        case blockTypes.HEADING_4:
-        case blockTypes.HEADING_5:
-        case blockTypes.HEADING_6: {
-          const level = getHeadingLevel(type);
-
-          if (
-            $from.node().type.name === 'heading' &&
-            $from.node().attrs.level === level
-          ) {
-            setProsemirrorBlockType(view.state.schema.nodes.paragraph)(
-              view.state,
-              view.dispatch,
-              view
-            );
-          } else {
-            setProsemirrorBlockType(view.state.schema.nodes.heading, {
-              level,
-            })(view.state, view.dispatch, view);
-          }
-          break;
-        }
-        case blockTypes.CODE_BLOCK:
-          setProsemirrorBlockType(view.state.schema.nodes.code_block)(
-            view.state,
-            view.dispatch,
-            view
-          );
-          break;
-        case blockTypes.BLOCKQUOTE:
-          wrapIn(view.state.schema.nodes.blockquote)(
-            view.state,
-            view.dispatch,
-            view
-          );
-          break;
-        case blockTypes.BULLET_LIST:
-          wrapInList(view.state.schema.nodes.bullet_list)(
-            view.state,
-            view.dispatch,
-            view
-          );
-          break;
-        case blockTypes.ORDERED_LIST:
-          wrapInList(view.state.schema.nodes.ordered_list)(
-            view.state,
-            view.dispatch,
-            view
-          );
-          break;
-        case blockTypes.PARAGRAPH:
-        default:
-          setProsemirrorBlockType(view.state.schema.nodes.paragraph)(
-            view.state,
-            view.dispatch,
-            view
-          );
-          break;
-      }
-    }
-  };
-  const handleStrongToggle = () => {
-    if (view && schema) {
-      toggleStrong(schema)(view.state, view.dispatch);
-      view.focus();
-    }
-  };
-
-  const handleEmToggle = () => {
-    if (view && schema) {
-      toggleEm(schema)(view.state, view.dispatch);
-      view.focus();
-    }
-  };
-
-  const handleLinkToggle = () => {
-    if (view && schema) {
-      if (!isMarkActive(schema.marks.link)(view.state)) {
-        const selectedText = getSelectedText(view.state);
-        setLinkDialogInitialAttrs({ title: selectedText ?? '', href: '' });
-        setIsLinkDialogOpen(true);
-      } else {
-        handleEditLink();
-      }
-
-      view.focus();
-    }
-  };
-
-  const handleCodeToggle = () => {
-    if (view && schema) {
-      toggleCode(schema)(view.state, view.dispatch);
-      view.focus();
-    }
-  };
-
-  const handleSaveLink = (attrs: LinkAttrs) => {
-    if (view && schema) {
-      if (!isMarkActive(schema.marks.link)(view.state)) {
-        addLink(schema)(attrs)(view.state, view.dispatch);
-      } else {
-        updateLink(schema)(attrs)(view.state, view.dispatch);
-      }
-      view.focus();
-    }
-
-    setIsLinkDialogOpen(false);
-  };
-
-  const handleEditLink = () => {
-    if (selectedLinkData) {
-      setLinkDialogInitialAttrs(selectedLinkData.linkAttrs);
-      setIsLinkPopoverOpen(false);
-      setIsLinkDialogOpen(true);
-    }
-  };
-
-  const handleRemoveLink = () => {
-    if (view && schema) {
-      removeLink(schema)(view.state, view.dispatch);
-      view.focus();
-    }
-
-    setIsLinkPopoverOpen(false);
-    setIsLinkDialogOpen(false);
-  };
-
-  const handleCloseLinkPopover = () => {
-    setIsLinkPopoverOpen(false);
-    view?.focus();
-  };
-
-  const handleNoteClick = () => {
-    if (view && schema) {
-      insertNote(view.state, view.dispatch);
-      view.focus();
-    }
-  };
-
-  const handleHorizontalRuleClick = () => {
-    if (view && schema) {
-      insertHorizontalRule(view.state, view.dispatch);
-      view.focus();
-    }
-  };
-
-  const handleImageClick = async () => {
-    if (!view) return;
-    try {
-      await pickAndInsertFigure(pickAsset)(view);
-      view.focus();
-    } catch (err) {
-      console.error(err);
-    }
-  };
-
-  return (
-    <>
-      <div
-        className="flex flex-auto p-4 outline-none"
-        onClick={() => editorViewRef.current?.focus()}
-      >
-        <div
-          className="editor flex-auto font-editor"
-          id="editor"
-          ref={editorRoot}
-        />
-      </div>
-
-      {isEditable && leafBlockType && (
-        <div
-          className={clsx(
-            'absolute self-center drop-shadow transition-bottom',
-            isToolbarOpen ? 'bottom-4' : '-bottom-12'
-          )}
-        >
-          <EditorToolbar
-            leafBlockType={leafBlockType}
-            containerBlockType={containerBlockType}
-            onBlockSelect={handleBlockSelect}
-            strongSelected={strongSelected}
-            emSelected={emSelected}
-            selectionIsLink={selectionIsLink}
-            codeSelected={codeSelected}
-            onStrongToggle={handleStrongToggle}
-            onEmToggle={handleEmToggle}
-            onLinkToggle={handleLinkToggle}
-            onCodeToggle={handleCodeToggle}
-            onNoteClick={handleNoteClick}
-            onHorizontalRuleClick={handleHorizontalRuleClick}
-            horizontalRuleEnabled={horizontalRuleEnabled}
-            onImageClick={handleImageClick}
-            imageEnabled={imageEnabled}
-          />
-        </div>
-      )}
-      {isEditable && (
-        <>
-          <LinkDialog
-            initialLinkAttrs={linkDialogInitialAttrs}
-            isOpen={isLinkDialogOpen}
-            onCancel={() => setIsLinkDialogOpen(false)}
-            onSave={handleSaveLink}
-          />
-          <LinkPopover
-            linkData={selectedLinkData}
-            isOpen={isLinkPopoverOpen}
-            onEditLink={handleEditLink}
-            onRemoveLink={handleRemoveLink}
-            onClose={handleCloseLinkPopover}
-          />
-        </>
-      )}
-    </>
-  );
-};
