+++ src/renderer/src/components/editing/SnapshotDocumentEditor.test.tsx
@@ -12,7 +12,7 @@ import {
   ProseMirrorContext,
   type ProseMirrorContextType,
 } from '../../../../modules/domain/rich-text/react/prosemirror-context';
-import { RichTextEditor } from './RichTextEditor';
+import { SnapshotDocumentEditor } from './SnapshotDocumentEditor';
 
 const { schema } = prosemirror;
 
@@ -27,7 +27,7 @@ const paragraph = (text: string): PMNode =>
     schema.node('paragraph', null, [schema.text(text)]),
   ]);
 
-describe('RichTextEditor', () => {
+describe('SnapshotDocumentEditor', () => {
   // A document swapped in while its predecessor is still converting must win,
   // even if the predecessor's conversion resolves first.
   it('builds from the latest document, not a superseded one', async () => {
@@ -40,15 +40,16 @@ describe('RichTextEditor', () => {
     // Only the conversion is exercised on this path; the rest goes unused.
     const context = {
       setView: () => {},
+      clearViewIfCurrent: () => {},
       convertToProseMirror,
       parseMarkdown: vi.fn(),
     } as unknown as ProseMirrorContextType;
 
     const editor = (doc: RichTextDocument) => (
       <ProseMirrorContext.Provider value={context}>
-        <RichTextEditor
+        <SnapshotDocumentEditor
           doc={doc}
-          isEditable={false}
+          onDocChange={async () => {}}
           pickAsset={async () => null}
           resolveAssetSrc={(src) => src}
         />
@@ -62,5 +63,41 @@ describe('RichTextEditor', () => {
 
     await waitFor(() => expect(container.textContent).toContain('from latest'));
     expect(container.textContent).not.toContain('from superseded');
+  });
+
+  // A document replaced after the view exists swaps state in place; the view
+  // survives the switch.
+  it('swaps in a document that arrives after the view was created', async () => {
+    // Deterministic on input (like the real converter), since conversion may
+    // legitimately run more than once per document.
+    const convertToProseMirror = vi.fn(
+      async ({ document }: { document: RichTextDocument }) =>
+        paragraph(`from ${document.content}`)
+    );
+
+    const context = {
+      setView: () => {},
+      clearViewIfCurrent: () => {},
+      convertToProseMirror,
+      parseMarkdown: vi.fn(),
+    } as unknown as ProseMirrorContextType;
+
+    const editor = (doc: RichTextDocument) => (
+      <ProseMirrorContext.Provider value={context}>
+        <SnapshotDocumentEditor
+          doc={doc}
+          onDocChange={async () => {}}
+          pickAsset={async () => null}
+          resolveAssetSrc={(src) => src}
+        />
+      </ProseMirrorContext.Provider>
+    );
+
+    const { container, rerender } = render(editor(markdownDocument('first')));
+    await waitFor(() => expect(container.textContent).toContain('from first'));
+
+    rerender(editor(markdownDocument('second')));
+    await waitFor(() => expect(container.textContent).toContain('from second'));
+    expect(container.textContent).not.toContain('from first');
   });
 });
